Moved unit tests to a subfolder.
