+++ Test/Unit/APIComponentDataTest.php
@@ -5,7 +5,7 @@
  * Contains APIComponentDataTest.
  */
 
-namespace DrupalCodeBuilder\Test;
+namespace DrupalCodeBuilder\Test\Unit;
 
 /**
  * Tests the AdminSettingsForm generator class.
