+++ Test/Unit/Basic7Test.php
@@ -5,7 +5,7 @@
  * Contains Basic7Test.
  */
 
-namespace DrupalCodeBuilder\Test;
+namespace DrupalCodeBuilder\Test\Unit;
 
 /**
  * Basic test class.
