+++ Test/Unit/ComponentRouterItem7Test.php
@@ -5,7 +5,7 @@
  * Contains ComponentRouterItem8Test.
  */
 
-namespace DrupalCodeBuilder\Test;
+namespace DrupalCodeBuilder\Test\Unit;
 
 /**
  * Tests for Router item component.
