+++ Test/Unit/ComponentThemeHook8Test.php
@@ -5,7 +5,7 @@
  * Contains ComponentThemeHook8Test.
  */
 
-namespace DrupalCodeBuilder\Test;
+namespace DrupalCodeBuilder\Test\Unit;
 
 /**
  * Tests the AdminSettingsForm generator class.
